requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
imageio-ffmpeg==0.4.9
openai>=2.16.0
